+++ libs/notification/src/lib/notification/components/item/item.module.ts
@@ -4,7 +4,6 @@ import { RouterModule } from '@angular/router';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { ItemComponent } from './item.component';
 import { ImageReferenceModule } from '@blockframes/ui/media/image-reference/image-reference.module';
-import { FilterByDateModule } from '@blockframes/utils/pipes/filter-by-date.pipe';
 
 import { MatListModule } from '@angular/material/list';
 import { MatButtonModule } from '@angular/material/button';
@@ -18,7 +17,6 @@ import { MatIconModule } from '@angular/material/icon';
     RouterModule,
     FlexLayoutModule,
     ImageReferenceModule,
-    FilterByDateModule,
     MatListModule,
     MatButtonModule,
     MatIconModule,
